Fixes various issues.
Mostly relating to unit conversions. The heater changes the format
for automatic mode based on the unit setting, so this must be stored and
normalized to celsius for display in HA.

I'm still looking for a solution to display fahrenheit if selected, as
forcing celsius is not ideal. However this does at least fix the control
bugs for now.